requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==1.3.0
streamlit==1.28.1
numpy==1.26.0
python-multipart==0.0.6
requests==2.31.0
setuptools==68.2.2
pytest==7.4.3
httpx==0.25.1
pydub==0.25.1
SpeechRecognition==3.10.0
gTTS==2.3.2
langchain==0.0.335
watchdog==6.0.0
streamlit-mic-recorder